Revert Vagrant #624 changes. Fixes #649 (#650)
* Revert Vagrant #624 changes. Fixes #649

This reverts commit b957777f5b5a6d82fe10bbbd6a295afc2c1489bf.
Revert "Remove Vagrant option from runner (#644)"

This reverts commit 91b31bcc66490833014d9182d274ce0bd4dc6e93.
Further minor changes. Fixes #649

* Clean up revert changes. Adds pipfile back in and removes vagrantfile

Pip docs back to normal


Delete Vagrantfile


Copy the old Pipfile.lock in


More cleanup after reverting

* add kubectl to docs

* refactor restart_pods to be <26 lines :muscle:

* python <3 whitespace

* newline

* Minor comment change to be more clear
