--- kubeconfig_aimmo.yml
@@ -1,17 +1,0 @@
-apiVersion: v1
-kind: Config
-users:
-- name: myself
-  user:
-     username: myself
-     password: secret
-clusters:
-- name: local
-  cluster:
-     server: http://localhost:8080
-contexts:
-- context:
-    cluster: local
-    user: myself
-  name: service-account-context
-current-context: service-account-context
